first pass of login/logout/register is semi-complete.
It works, but the code is so awful, it's probably criminal (as in it's taken from a working example and munged until it fits).

do not use.
